split into two scripts so update can be used for general updates
